feat: Add optional checksums generation for each release.
Also refactor release providers into a seperate mods.

+++ src/lib.rs
@@ -1,18 +1,19 @@
-use crate::utils::get_latest_tag;
+use crate::{checksum::get_new_checksummer, utils::get_latest_tag};
 use camino::Utf8Path;
 use eyre::{bail, Context, Result};
 use log::{debug, error, info, warn};
+use release_provider::{docker, github::Github};
 use std::{env, sync::Arc};
+use tokio::io::AsyncWriteExt;
 use tokio::{fs, process::Command, sync::Mutex};
 
+mod checksum;
 pub mod config;
-mod docker;
-mod github;
-pub mod release_provider;
+mod release_provider;
 mod utils;
+
 use crate::release_provider::ReleaseProvider;
 use config::{Build, Config, Release};
-use github::Github;
 use utils::archive_file;
 
 #[derive(Debug, Clone)]
@@ -52,6 +53,12 @@ pub async fn run(cfg: Config, opts: Opts) -> Result<()> {
         // Wait until all builds are finished in a release.
         futures::future::join_all(&mut all_builds).await;
 
+        let rls = &releases[i];
+
+        if rls.checksum.is_some() {
+            create_checksums(rls, all_archives.clone()).await?;
+        }
+
         debug!("all archives generated: {:?}", all_archives);
         if opts.publish {
             let latest_tag = match get_latest_tag().await {
@@ -89,7 +96,7 @@ fn get_release_providers(release: &Release) -> Result<Vec<Box<dyn ReleaseProvide
 
     // Check if github details are provided.
     if release.targets.github.is_some() {
-        let ghtoken = get_github_token()?;
+        let ghtoken = get_github_token();
         let gh = Github::new(ghtoken);
         providers.push(Box::new(gh));
     }
@@ -110,16 +117,12 @@ pub async fn run_build(release: &Release, build: &Build, rm_dist: bool) -> Resul
     if output.status.success() {
         // Delete the dist directory if rm_dist is provided.
         if rm_dist {
-            // Check if the dist directory exists.
-            match fs::metadata(&release.dist_folder).await {
-                Ok(meta) => {
-                    if meta.is_dir() {
-                        fs::remove_dir_all(&release.dist_folder).await?;
-                    } else {
-                        bail!("error deleting dist dir: not a directory");
-                    }
+            if let Ok(meta) = fs::metadata(&release.dist_folder).await {
+                if meta.is_dir() {
+                    fs::remove_dir_all(&release.dist_folder).await?;
+                } else {
+                    bail!("error deleting dist dir: not a directory");
                 }
-                _ => {}
             }
         }
 
@@ -165,10 +168,43 @@ pub async fn run_build(release: &Release, build: &Build, rm_dist: bool) -> Resul
     Ok(String::from(""))
 }
 
-fn get_github_token() -> Result<String> {
+fn get_github_token() -> String {
     // Check if `GITHUB_TOKEN` is present.
-    match env::var("GITHUB_TOKEN") {
-        Ok(token) => Ok(token),
-        Err(_) => Ok(String::from("")),
+    env::var("GITHUB_TOKEN").unwrap_or_else(|_| String::new())
+}
+
+async fn create_checksums(rls: &Release, all_archives: Arc<Mutex<Vec<String>>>) -> Result<()> {
+    let cm_path = Utf8Path::new(&rls.dist_folder).join("checksums.txt");
+    if let Ok(_) = fs::metadata(&cm_path).await {
+        // Remove checksums file if it exists.
+        fs::remove_file(&cm_path).await?;
     }
+
+    // Open the file with options set to both create (if it doesn't exist) and append
+    let mut file = fs::OpenOptions::new()
+        .create(true) // create if it doesn't exist
+        .append(true) // set to append mode
+        .open(&cm_path)
+        .await?;
+    let archives = all_archives.lock().await.clone();
+    for arc in archives {
+        let path = Utf8Path::new(&arc);
+
+        let cm = get_new_checksummer(rls.checksum.as_ref().unwrap().algorithm.as_ref())?;
+
+        let checksum = cm.compute(&arc).await?;
+
+        debug!(
+            "writing to checksums file: {}, {}",
+            path.file_name().unwrap(),
+            &checksum
+        );
+        // Write the name and checksum to the file
+        file.write(format!("{}\t{}\n", path.file_name().unwrap(), checksum).as_bytes())
+            .await?;
+
+        file.flush().await?;
+    }
+
+    Ok(())
 }
